delete: CommunicationFaultData CommunicationFaultObjectPair SimulationFault Moved EventCommand and FaultCommand classes to dto.events.
