+++ gov.pnnl.goss.gridappsd/src/gov/pnnl/goss/gridappsd/dto/events/EventCommand.java
@@ -1,11 +1,11 @@
-package gov.pnnl.goss.gridappsd.dto;
+package gov.pnnl.goss.gridappsd.dto.events;
 
 import java.io.Serializable;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonElement;
 
-import gov.pnnl.goss.gridappsd.dto.events.CommOutage;
+import gov.pnnl.goss.gridappsd.dto.BaseEventCommand;
 
 public class EventCommand extends BaseEventCommand implements Serializable {
 	/**
