+++ gov.pnnl.goss.gridappsd/test/gov/pnnl/goss/gridappsd/TestCommunicationFault.java
@@ -9,13 +9,13 @@ import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonObject;
 
-import gov.pnnl.goss.gridappsd.dto.CommunicationFault;
-import gov.pnnl.goss.gridappsd.dto.CommunicationFaultData;
-import gov.pnnl.goss.gridappsd.dto.CommunicationFaultObjectPair;
+//import gov.pnnl.goss.gridappsd.dto.CommunicationFault;
+//import gov.pnnl.goss.gridappsd.dto.CommunicationFaultData;
+//import gov.pnnl.goss.gridappsd.dto.CommunicationFaultObjectPair;
 import gov.pnnl.goss.gridappsd.dto.DifferenceMessage;
-import gov.pnnl.goss.gridappsd.dto.EventCommand;
-import gov.pnnl.goss.gridappsd.dto.FaultCommand;
+import gov.pnnl.goss.gridappsd.dto.events.EventCommand;
 import gov.pnnl.goss.gridappsd.dto.events.Fault;
+import gov.pnnl.goss.gridappsd.dto.events.FaultCommand;
 
 
 public class TestCommunicationFault {
@@ -40,28 +40,28 @@ public class TestCommunicationFault {
 		DifferenceMessage dm = new DifferenceMessage ();
 		dm.difference_mrid="1234";
 
-		CommunicationFault commFault = new CommunicationFault();
-		commFault.object = "UU1234";
-		commFault.attribute = "FilterObject";
-		commFault.value = new CommunicationFaultData();
-		CommunicationFaultObjectPair tempCFOP = new CommunicationFaultObjectPair();
-		tempCFOP.objectMRID = "UU123214";
-		tempCFOP.attribute = "RegulatingControl.mode";
-		commFault.value.inputList.add(tempCFOP); 
-		commFault.value.outputList.add("UU12323");
-		commFault.value.filterAllInputs = false;
-		commFault.value.filterAllOutputs = false;
-		commFault.value.occuredDateTime = 1248156005L;
-		commFault.value.stopDateTime = 1248156008L;
+//		CommunicationFault commFault = new CommunicationFault();
+//		commFault.object = "UU1234";
+//		commFault.attribute = "FilterObject";
+//		commFault.value = new CommunicationFaultData();
+//		CommunicationFaultObjectPair tempCFOP = new CommunicationFaultObjectPair();
+//		tempCFOP.objectMRID = "UU123214";
+//		tempCFOP.attribute = "RegulatingControl.mode";
+//		commFault.value.inputList.add(tempCFOP); 
+//		commFault.value.outputList.add("UU12323");
+//		commFault.value.filterAllInputs = false;
+//		commFault.value.filterAllOutputs = false;
+//		commFault.value.occuredDateTime = 1248156005L;
+//		commFault.value.stopDateTime = 1248156008L;
 
 		
-		System.out.println(new String (commFault.toString()));
-		String tempStr = "{\"inputList\":[{\"objectMRID\":\"UU123214\",\"attribute\":\"RegulatingControl.mode\"}],\"outputList\":[\"UU12323\"],\"filterAllInputs\":false,\"filterAllOutputs\":false,\"timeInitiated\":1248156005,\"timeCleared\":1248156008}";
-		CommunicationFaultData tempComm = CommunicationFaultData.parse(tempStr);
-		System.out.println(gson.toJson(tempComm));
-		System.out.println("************************");
-		System.out.println(gson.toJson(tempComm.buildSimFault()));
-		System.out.println("************************");
+//		System.out.println(new String (commFault.toString()));
+//		String tempStr = "{\"inputList\":[{\"objectMRID\":\"UU123214\",\"attribute\":\"RegulatingControl.mode\"}],\"outputList\":[\"UU12323\"],\"filterAllInputs\":false,\"filterAllOutputs\":false,\"timeInitiated\":1248156005,\"timeCleared\":1248156008}";
+//		CommunicationFaultData tempComm = CommunicationFaultData.parse(tempStr);
+//		System.out.println(gson.toJson(tempComm));
+//		System.out.println("************************");
+//		System.out.println(gson.toJson(tempComm.buildSimFault()));
+//		System.out.println("************************");
 		
 //		EventCommand eventCommand = new EventCommand();
 //		eventCommand.command = "CommEvent";
@@ -98,18 +98,18 @@ public class TestCommunicationFault {
 //		BaseEventCommand testEventCommand2 = gson.fromJson(tempString, BaseEventCommand.class);
 //		System.out.println(gson.toJson(testEventCommand2));
 		
-		System.out.println(gson.toJson(commFault));
-		
-		dm.forward_differences.clear();
-		dm.reverse_differences = null;
-		dm.forward_differences.add(commFault);
-		System.out.println(dm.toString());
-		
-		dm = new DifferenceMessage ();
-		dm.difference_mrid="_"+UUID.randomUUID();
-		dm.forward_differences.add(commFault);
-//		dm.reverse_differences.add();
-		System.out.println(dm.toString());
+//		System.out.println(gson.toJson(commFault));
+//		
+//		dm.forward_differences.clear();
+//		dm.reverse_differences = null;
+//		dm.forward_differences.add(commFault);
+//		System.out.println(dm.toString());
+//		
+//		dm = new DifferenceMessage ();
+//		dm.difference_mrid="_"+UUID.randomUUID();
+//		dm.forward_differences.add(commFault);
+////		dm.reverse_differences.add();
+//		System.out.println(dm.toString());
 		
 		JsonObject input = new JsonObject();
 		input.addProperty("simulation_id", 1231234567);
