+++ gov.pnnl.goss.gridappsd/test/gov/pnnl/goss/gridappsd/TestFaultMessage.java
@@ -11,7 +11,7 @@ import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 
 import gov.pnnl.goss.gridappsd.dto.DifferenceMessage;
-import gov.pnnl.goss.gridappsd.dto.events.SimulationFault;
+//import gov.pnnl.goss.gridappsd.dto.events.SimulationFault;
 
 
 
@@ -23,15 +23,15 @@ public class TestFaultMessage {
 		DifferenceMessage dm = new DifferenceMessage ();
 		dm.difference_mrid="_"+UUID.randomUUID();
 
-		SimulationFault simFault = new SimulationFault();
-		simFault.eventId = "_1f4467ee-678b-49c6-b58c-9f9462cf5ae4";
+//		SimulationFault simFault = new SimulationFault();
+//		simFault.eventId = "_1f4467ee-678b-49c6-b58c-9f9462cf5ae4";
 //		simFault.FaultImpedance = new FaultImpedance();
 //		simFault.FaultImpedance.rGround = 0.0;
 //		simFault.FaultImpedance.xGround = 0.5;
 //		simFault.PhaseConnectedFaultKind="lineToGround";
 		
 		JsonArray faults = new JsonArray();
-		faults.add(simFault.toJsonElement()); 
+//		faults.add(simFault.toJsonElement()); 
 		JsonObject topElement = new JsonObject();
 		topElement.add("Faults", faults);
 //		System.out.println(simFault.toString());
@@ -39,7 +39,7 @@ public class TestFaultMessage {
 		
 		dm.forward_differences.clear();
 		dm.reverse_differences = null;
-		dm.forward_differences.add(simFault);
+//		dm.forward_differences.add(simFault);
 		System.out.println(gson.toJson(dm));
 		
 		dm = new DifferenceMessage ();
